corrected page rank teleport

+++ src/main/java/org/nonari/fic/ri/utils/page_rank/PageRank.java
@@ -38,7 +38,7 @@ public class PageRank {
             }
 
             if (c.getComm().equals("teleport")){
-                teleport = (Integer)c.getArgs().get(0);
+                teleport = (Float) c.getArgs().get(0);
             }
 
             if (c.getComm().equals("calc")){
